+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Query a single term in a Managed Resource
 
 ### Fixed
+- Syntax error in request with facet queries that contain local parameters
 - HEAD requests could lead to timeouts with cURL adapter
 
 ### Changed